Create CODEOWNERS
We want the branch protection rule that all changes has to happen via pull request.
This was not possible when only Alex had write access. (A review was required by someone else for him to make changes)
It has been decided that for now, Coralie will approve all pull requests. 
This CODEOWNERS file should automatically assign Coralie as a reviewer of all Pull-Requests.